Update to only run tasks based on whats actaully there.
